Move DeviceUsage to mfa/types.
