+++ web/packages/teleport/src/Account/ManageDevices/useManageDevices.ts
@@ -21,8 +21,8 @@ import useAttempt from 'shared/hooks/useAttemptNext';
 
 import Ctx from 'teleport/teleportContext';
 import cfg from 'teleport/config';
-import auth, { DeviceUsage } from 'teleport/services/auth';
-import { MfaDevice } from 'teleport/services/mfa';
+import auth from 'teleport/services/auth';
+import { DeviceUsage, MfaDevice } from 'teleport/services/mfa';
 import { MfaChallengeScope } from 'teleport/services/auth/auth';
 
 export default function useManageDevices(ctx: Ctx) {
